Создан класс Game
